backend/requirements.txt
kalshi-python>=2.0.0
websockets>=12.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
cryptography>=41.0.0
aiohttp>=3.9.0
